requirements.txt
python-twitter==2.3
supervisor>=3.0a8
RPi.GPIO>=0.5.11